surveys and applications
