+++ src/main/resources/templates/applications-and-needs package/applications-and-needs.html
@@ -13,7 +13,8 @@
         <!--Section heading-->
         <h2 class="h1-responsive font-weight-bold text-center my-4">Masz pytanie? Napisz!</h2>
         <!--Section description-->
-        <p class="text-center w-responsive mx-auto mb-5">Jeśli masz jakieś pytania, jesteśmy tu aby na nie odpowiedzieć.</p>
+        <p class="text-center w-responsive mx-auto mb-5">Jeśli masz jakieś pytania, jesteśmy tu aby na nie
+            odpowiedzieć.</p>
 
         <div class="row">
 
@@ -22,19 +23,15 @@
                 <form id="contact-form" name="contact-form" action="mail.php" method="POST">
 
                     <!--Grid row-->
-                    <div class="row">
-
+                    <div class="row" th:if="${!#request.isUserInRole('ADMIN') and !#request.isUserInRole('USER')}">
                         <!--Grid column-->
                         <div class="col-md-6">
                             <div class="md-form mb-0">
                                 <input type="text" id="name" name="name" class="form-control">
                                 <label for="name" class="">Twoje imię</label>
-
-
                             </div>
                         </div>
                         <!--Grid column-->
-
                         <!--Grid column-->
                         <div class="col-md-6">
                             <div class="md-form mb-0">
@@ -42,10 +39,23 @@
                                 <label for="email" class="">Twój email</label>
                             </div>
                         </div>
-                        <!--Grid column-->
 
+
+                        <!--Grid column-->
                     </div>
                     <!--Grid row-->
+
+                    <div class="row" th:if="${#request.isUserInRole('ADMIN') or #request.isUserInRole('USER')}">
+                        <div class="col-md-6">
+                            <div class="md-form mb-0">
+                                <input type="text" id="email" th:value="${user_email}" name="email" disabled hidden
+                                       class="form-control">
+<!--                                <label for="email" class="">Twój email</label>-->
+                            </div>
+                        </div>
+                    </div>
+
+
 
                     <!--Grid row-->
                     <div class="row">
@@ -65,7 +75,8 @@
                         <div class="col-md-12">
 
                             <div class="md-form">
-                                <textarea type="text" id="message" name="message" rows="2" class="form-control md-textarea"></textarea>
+                                <textarea type="text" id="message" name="message" rows="2"
+                                          class="form-control md-textarea"></textarea>
                                 <label for="message">Twoja wiadomość</label>
                             </div>
 
@@ -76,7 +87,8 @@
                 </form>
 
                 <div class="text-center text-md-left">
-                    <a class="btn btn-primary" onclick="document.getElementById('contact-form').submit();" style="color: white">Wyślij</a>
+                    <a class="btn btn-primary" onclick="document.getElementById('contact-form').submit();"
+                       style="color: white">Wyślij</a>
                 </div>
                 <div class="status"></div>
             </div>
